✨ Refactor map components: Introduce MapContainer, RouteLine, RouteMarkers, and control components for improved structure and maintainability
