+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/components/map.tsx","./src/components/mapcontrols.tsx","./src/components/routemarker.tsx","./src/constants/map.ts","./src/hooks/usemaphandlers.ts","./src/hooks/usemarkerdrag.ts","./src/store/routestore.ts","./src/store/uistore.ts","./src/types/index.ts","./src/utils/geo.ts","./src/utils/gpx.ts","./src/utils/maphelpers.ts"],"version":"5.7.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/components/map.tsx","./src/components/mapcontrols.tsx","./src/components/routemarker.tsx","./src/components/controls/editcontrols.tsx","./src/components/controls/historycontrols.tsx","./src/components/controls/modeindicator.tsx","./src/components/controls/routeactions.tsx","./src/components/controls/routeinfobar.tsx","./src/constants/map.ts","./src/hooks/usemaphandlers.ts","./src/hooks/usemarkerdrag.ts","./src/store/routestore.ts","./src/store/uistore.ts","./src/types/index.ts","./src/utils/geo.ts","./src/utils/gpx.ts","./src/utils/maphelpers.ts"],"version":"5.7.3"}
